ui update
latest ui update
(not done with feedback)
(dictionary design)
