+++ styles/style4.css
@@ -53,6 +53,7 @@ nav .logo span {
   transform: translateX(100%);
   z-index: 1;
   margin-top: 65px;
+  display: none;
 }
 
 .sidebar.open {
@@ -108,6 +109,12 @@ nav .logo span {
   background-color: #fffafa;
 }
 
+textarea {
+  border: none;
+  outline: none;
+  font-size: 24px;
+  font-family: "Poppins", sans-serif;
+}
 #error-message {
   display: none;
 }
@@ -115,6 +122,7 @@ nav .logo span {
   display: flex;
   flex-direction: row;
   width: 80%;
+  z-index: 1;
 }
 .box {
   position: relative;
@@ -128,6 +136,7 @@ nav .logo span {
   width: 45vw; /* Adjust width for larger screens */
   border-top-left-radius: 10px;
   border-bottom-left-radius: 10px;
+  z-index: 0;
 }
 
 .box2 {
@@ -143,6 +152,7 @@ nav .logo span {
   width: 45vw; /* Adjust width for larger screens */
   border-top-right-radius: 10px;
   border-bottom-right-radius: 10px;
+  z-index: 0;
 }
 
 .arrow-down {
@@ -167,15 +177,6 @@ nav .logo span {
   width: 100px; /* Adjust the width as needed */
   margin-bottom: 10px;
 }
-.details {
-  display: flex;
-  width: 100%;
-  position: absolute;
-  bottom: 10px;
-  left: 10px;
-  background-color: white;
-  color: white;
-}
 
 .upload-file {
   border: 2px dashed #ccc; /* Initial border style */
@@ -186,16 +187,6 @@ nav .logo span {
   z-index: 1;
 }
 
-.upload-file.dragover {
-  border: 2px dashed #87ceeb; /* Change border color when dragging over */
-  background-color: rgba(
-    135,
-    206,
-    235,
-    0.1
-  ); /* Optional: Light up the background */
-  box-shadow: 0px 0px 10px rgba(135, 206, 235, 0.5); /* Optional: Add box shadow */
-}
 .translate-button {
   position: absolute;
   bottom: 10px;
@@ -248,7 +239,6 @@ nav .logo span {
   justify-content: flex-end;
   width: 100%;
   padding: 5px;
-  display: none;
 }
 .header-downloadfile button {
   background-color: #fffafa;
@@ -273,7 +263,7 @@ nav .logo span {
   display: flex;
   width: 50%;
 
-  padding: 5px;
+  padding: 20px;
   justify-content: flex-start;
 }
 .language-selector {
@@ -286,6 +276,9 @@ nav .logo span {
 }
 .language-selector .label1 {
   font-weight: 700;
+  margin: 0;
+  font-size: 20px;
+  color: #333333;
 }
 
 .language-selector span:hover {
@@ -297,18 +290,6 @@ nav .logo span {
 .language-selector span.active {
   text-decoration: underline; /* Add underline to clicked spans */
 }
-.language-selector img {
-  width: 30px;
-  height: 30px;
-  margin-right: 10px;
-}
-.language-selector label {
-  margin: 0;
-  font-size: 16px;
-
-  color: #333333;
-}
-
 .upload-icontext {
   position: absolute;
   bottom: 10px;
@@ -363,16 +344,148 @@ nav .logo span {
   font-family: "Poppins", sans-serif;
   font-size: 0.8em;
 }
+.textinput {
+  border: none;
+  resize: none;
+  width: 100%;
+  min-height: 200px;
+}
 
 .custom-textfield {
   resize: none;
   border-radius: 10px;
   font-size: 16px;
-  background-color: var(--light);
-
-  width: 500px;
-  height: 100px; /* Adjust the height as needed */
+  width: calc(
+    100% - 40px
+  ); /* Subtract any padding/margin from the parent container */
+  height: 100%;
   vertical-align: top;
-  overflow-y: auto; /* Add scrollbars when content exceeds the height */
-  color: #181a1e;
+  padding: 5px; /* Add padding if needed */
+}
+
+.output-p {
+  width: 100%;
+  height: 100%;
+}
+
+.output-p p {
+  border: none !important;
+  outline: none !important;
+  font-size: 24px;
+  font-family: "Poppins", sans-serif;
+}
+
+.container.drag-hover > .box,
+.container.drag-hover > .box2,
+.container.drag-hover > .box > .details {
+  background-color: #e6effe; /* Or your desired highlight color */
+  border: 2px solid #3080ea;
+  border-radius: 10px; /* Adjust as needed */
+}
+
+.drop-text {
+  /* Set desired text properties like color, font, and size */
+  color: #3080ea; /* White color */
+  font-size: 100px;
+  font-weight: 400;
+  text-align: center; /* Center the text horizontally */
+  position: absolute; /* Position it relative to the container */
+  top: 50%; /* Position it vertically at 50% */
+  left: 50%; /* Position it horizontally at 50% */
+  transform: translate(-50%, -50%); /* Adjust positioning as needed */
+  opacity: 0; /* Initially hide the text */
+}
+
+.container.drag-hover .drop-text {
+  opacity: 1; /* Show the text on hover */
+}
+
+.text-section {
+  display: flex;
+  width: 100%;
+  flex-direction: column;
+  min-height: 300px;
+}
+
+.text-section .customtextfield {
+  height: 500px;
+}
+
+.details {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+.details button {
+  appearance: none;
+  border: none;
+  outline: none;
+  cursor: pointer;
+  background: none;
+  color: crimson;
+}
+.modalloading {
+  display: none;
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(0, 0, 0, 0.5); /* Make the entire modal transparent */
+  justify-content: center;
+  align-items: center;
+  z-index: 2;
+  border: none; /* Remove border for the modal */
+}
+
+.modal-contentloading {
+  background-color: transparent; /* Make the content background transparent */
+  border: none; /* Remove border for the content */
+  text-align: center;
+  z-index: 3;
+}
+
+/* Ensure the image is centered */
+.modal-contentloading img {
+  display: block;
+  margin: 0 auto;
+}
+.modal-contentloading p {
+  color: white;
+  font-size: 26px;
+}
+
+.dict-div {
+  width: 50%;
+  display: flex;
+  flex-direction: column;
+}
+
+.options {
+  display: none;
+}
+
+.feedback {
+  display: flex;
+  width: 80%;
+  justify-content: flex-end;
+  flex-direction: row;
+}
+
+.tryagain {
+  margin-top: 15px;
+  background-color: #87ceeb;
+  color: white;
+  padding: 10px 20px;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+  font-size: 24px;
+  position: fixed;
+  font-family: "Poppins", sans-serif;
+  position: absolute;
+  margin-right: 200px;
+  right: 0;
 }
