+++ text-text.php
@@ -13,7 +13,7 @@ require("utilities/recent_text_translation.php");
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link href='https://unpkg.com/boxicons@2.1.4/css/boxicons.min.css' rel='stylesheet'>
     <!-- <link rel="stylesheet" href="styles/style2.css"> -->
-    <link rel="stylesheet" href="styles/style2.css">
+    <link rel="stylesheet" href="styles/style4.css">
     <link rel="stylesheet" href="styles/simplePagination.css">
     <title>Text to Text Translation</title>
     <link rel="icon" type="image/x-icon" href="images/icon.ico">
@@ -36,88 +36,89 @@ require("utilities/recent_text_translation.php");
 
 <body>
     <!-- Main Content -->
-    <div class="content">
-        <!-- Navbar -->
-        <nav>
-            <i class='bx bx-menu'></i>
-        </nav>
 
-        <!-- End of Navbar -->
-
-        <main>
-
-            <!-- Add this div for the loading feature -->
-            <!-- Loading Modal after clicking translate button -->
-
-            <div class="header">
-                <div class="left">
-                    <h1>
-                        Text-Text Translator
-                    </h1>
-
-
-                    <!-- Error Message: Pabago nalang if may naiisip kang ibang design -->
-                    <p style="color: red;" id="error-message"><i>
-                            <?php
-                            if (isset($_GET['error'])) {
-                                switch ($_GET['error']) {
-                                    case 1: // user did not chooose language
-                                        echo "Please select a source/translated language.";
-                                        break;
-                                    case 2: // user did not enter text
-                                        echo "Please choose two different language.";
-                                        break;
-                                    case 3: // user chose two same language
-                                        echo "Please type text to be translated.";
-                                        break;
-                                    case 4: // user added unprovided choices
-                                        echo "Please choose only on the provided models/languages.";
-                                        break;
-                                }
-                            }
-                            ?>
-                        </i></p>
-
-
-                </div>
+    <!-- Navbar -->
+    <nav>
+        <div class="logo">
+            <img src="images/logo.png" alt="Voce logo">
+            <span>Text Translator</span>
+        </div>
+    </nav>
+    <!-- End of Navbar -->
+    <div class="sidebar">
+        <div class="top-iconsheader">
+            <div class="top-icons">
+                <a href=""><img src="images/anti-clockwise.png" alt="Language Icon" width="30px"></a>
+                <a href=""><img src="images/anti-clockwise.png" alt="Language Icon" width="30px"></a>
             </div>
-
-
-
-
-            <!-- Truncate Text -->
-            <div id="myModal" class="modal">
-                <div class="modal-content">
-                    <span class="close">&times;</span>
-                    <p id="modalText"></p>
-                </div>
+            <div class="top-close">
+                <a href=""><img src="images/xmas.png" alt="Language Icon" width="30px"></a>
             </div>
-
-            <div class="bottom-data">
-
-                <!-- SOURCE LANGUAGE DIVISION-->
-                <div class="translator">
-                    <div class="header">
-                        <h3>Source Language</h3>
-                        <br>
-                    </div>
-
-                    <div id="loadingModal" class="modalloading">
-                        <div class="modal-contentloading">
-                            <img src="images/loading.gif" alt="Loading..." />
-                            <p>Loading....</p>
-                        </div>
-                    </div>
-
-                    <!-- START OF FORM, COVERS TWO SELECT AND ONE TEXT AREA -->
+        </div>
+        <div class="sidebar-header">
+            <h2 style="margin-left: 10px;">History</h2>
+        </div>
+        <div class="sidebar-headerclear">
+            <a href="#" style="margin-right: 4px; padding: 5px;">Clear all history</a>
+        </div>
+        <br>
+        <div class="sidebar-history">
+            Testting first
+            Testting first
+            Testting first
+            Testting first
+        </div>
+        <div class="robot-section">
+            <img src="images/hi-robot.gif" class="robot" alt="robot waving" width="80%">
+        </div>
+    </div>
+    <div class="main-content">
+        <div class="header">
+            <a href="text-text.php"><button><img src="images/translator.png" alt="Language Icon"
+                        width="30px">Text</button></a>
+            <a href="index.php"><button><img src="images/music-file.png" alt="Language Icon" width="30px">Upload
+                    a File</button></a>
+        </div>
+        <div class="header-downloadfile-wrapper">
+            <div class="header-downloadfile" dir="rtl" id="download-file" style="display: none;">
+                <form method="post" action="utilities/download_text_output.php">
+                    <button type="submit" name="text">
+                        <img src="images/write.png" alt="Language Icon" width="30px">Download as Text File</button>
+                    <button type="submit" name="word"><img src="images/download.png" alt="Language Icon"
+                            width="30px">Download as Word File</button>
+                </form>
+            </div>
+        </div>
+        <!-- Error Message: Pabago nalang if may naiisip kang ibang design -->
+        <p style="color: red;" id="error-message"><i>
+                <?php
+                if (isset($_GET['error'])) {
+                    switch ($_GET['error']) {
+                        case 1: // user did not chooose language
+                            echo "Please select a source/translated language.";
+                            break;
+                        case 2: // user did not enter text
+                            echo "Please choose two different language.";
+                            break;
+                        case 3: // user chose two same language
+                            echo "Please type text to be translated.";
+                            break;
+                        case 4: // user added unprovided choices
+                            echo "Please choose only on the provided models/languages.";
+                            break;
+                    }
+                }
+                ?>
+            </i></p>
+        <div class="header-language selector">
+            <!-- SELECT LANGUAGE -->
+            <div class="language-selector-wrapper">
+                <div class="language-selector">
+                    <label class="label1">
+                        Source language:
+                    </label>
                     <form id="myForm" action="utilities/text_translation.php" method="POST" onsubmit="showLoading()">
-                        <?php // url must have translated=1 before showing the output
-                        if (isset($_SESSION['recent_text'])) {
-                            $textid = $_SESSION['recent_text'];
-                            $data = mysqli_query($dbcon, "SELECT * FROM text_translations WHERE text_id = '$textid' AND from_audio_file = 0 ORDER BY translation_date DESC LIMIT 1")->fetch_row();
-                        }
-                        ?>
-                        <!-- SELECT LANGUAGE -->
+
                         <select name="src" id="sourceLanguage">
                             <option value="">Select One …</option>
                             <?php foreach ($common_langs as $lang => $code): ?>
@@ -126,37 +127,15 @@ require("utilities/recent_text_translation.php");
                                 </option>
                             <?php endforeach ?>
                         </select>
-                        <br>
-                        <?php if (isset($_SESSION['recent_text']) && isset($_GET['translated']) && $_GET['translated'] == 1) {
-                            echo "Language: " . $data[4];
-                        } ?>
-                        <!-- <input type = "text" name = "text" class="form-control"> -->
-                        <textarea class="custom-textfield" name="text" id="text-input" placeholder='Type Here...'><?php
-                        if (isset($_SESSION['recent_text']) && isset($_GET['translated']) && $_GET['translated'] == 1) {
-                            echo $data[6] ?? '';
-                        }
-                        ?></textarea>
-
-                        <!-- Button to submit -->
-                        <!--<button type = "submit" id="yourButtonID" class="custom-button">Translate</button> -->
-
-                        <a href="javascript:void(0);"
-                            onclick="document.getElementById('myForm').submit(); showLoading();">
-                            <img src="images/arrows.png" alt="Image Description" class="center-image" width="50px"
-                                height="40px">
-                        </a>
-
                 </div>
-
-                <!-- TARGETED LANGUAGE DIVISION -->
-                <div class="translator">
-                    <div class="header">
-                        <h3>Targeted Language</h3>
-                        <br>
-                    </div>
-
-                    <!-- SELECT LANGUAGE -->
-                    <select name="target" class="form-control" id="targetLanguage">
+            </div>
+            <div class="language-selector-wrapper">
+                <div class="language-selector">
+                    <label class="label1">
+                        Target language:
+                    </label>
+                    <select name="target" id="targetLanguage" class="form-control">
+                        <!-- Will display languages supported by API only-->
                         <option value="">Select One …</option>
                         <?php foreach ($common_langs as $lang => $code): ?>
                             <option name="language">
@@ -164,41 +143,77 @@ require("utilities/recent_text_translation.php");
                             </option>
                         <?php endforeach ?>
                     </select>
-                    <br>
+
+                </div>
+            </div>
+        </div>
+        <div class="container">
+            <div class="box">
+                <!-- START OF FORM, COVERS TWO SELECT AND ONE TEXT AREA -->
+                <div class="custom-textfield">
+
+                    <?php // url must have translated=1 before showing the output
+                    if (isset($_SESSION['recent_text'])) {
+                        $textid = $_SESSION['recent_text'];
+                        $data = mysqli_query($dbcon, "SELECT * FROM text_translations WHERE text_id = '$textid' AND from_audio_file = 0 ORDER BY translation_date DESC LIMIT 1")->fetch_row();
+                    }
+                    ?>
+                    <?php if (isset($_SESSION['recent_text']) && isset($_GET['translated']) && $_GET['translated'] == 1) {
+                        echo "Language: " . $data[4];
+                    } ?>
+                    <textarea class="textinput" name="text" id="text-input" placeholder='Type Here...'><?php
+                    if (isset($_SESSION['recent_text']) && isset($_GET['translated']) && $_GET['translated'] == 1) {
+                        echo $data[6] ?? '';
+                    }
+                    ?></textarea>
+
                     <?php if (isset($_SESSION['recent_text']) && isset($_GET['translated']) && $_GET['translated'] == 1) {
                         echo "Language: " . $data[5];
                     } ?>
-                    <!-- END OF FORM -->
-                    </form>
-
-
-
-                    <!-- Output text-->
-                    <div class="custom-textfield" contenteditable="true" readonly>
-                        <p class="test outputText" id="text-output">
-                        </p>
-                    </div>
-                </div>
-                <div class="dict-div">
-                    <h2 class="hovered-word">Word</h2>
-                    <p class="word-meaning">Meaning: </p>
-
-
-                </div>
-
-                <div class="download button" dir="rtl" id="download-file" style="display:none;">
-                    <form method="post" action="utilities/download_text_output.php">
-                        <button type="submit" name="text" style="padding:5px;">Download as Text File</button>
-                        <button type="submit" name="word" style="padding:5px;">Download as Word File</button>
+                    <!-- end of form -->
                     </form>
                 </div>
             </div>
-        </main>
-
+            <div class="box2">
+                <!-- Output text-->
+                <div class="output-p" contenteditable="true" readonly>
+                    <p class="test outputText" id="text-output">
+                    </p>
+                </div>
+            </div>
+            <div class="options">
+                <a href=""><img src="images/anti-clockwise.png" alt="Language Icon" width="30px"></a>
+                <label>History</label>
+                <br>
+                <a href=""><img src="images/dictionary-icon.png" alt="Language Icon" width="30px"></a>
+                <label>Dictionary</label>
+            </div>
+        </div>
+        <div class="feedback">
+            <a href="feedback">
+                <p> send feedback </p>
+            </a>
+        </div>
+        <!-- dictionary-->
+        <div class="dict-div">
+            <h2 class="hovered-word">Word</h2>
+            <p class="word-meaning">Meaning: </p>
+        </div>
 
     </div>
+    <div class="download button" dir="rtl" id="download-file" style="display:none;">
+        <form method="post" action="utilities/download_text_output.php">
+            <button type="submit" name="text" style="padding:5px;">Download as Text File</button>
+            <button type="submit" name="word" style="padding:5px;">Download as Word File</button>
+        </form>
+    </div>
+    </div>
 
-    <!-- for an in-depth walkthrough for pagination, please visit https://bilalakil.me/simplepagination -->
+
+
+
+
+
     <script src="scripts/dictionary.js"></script>
     <script src="https://code.jquery.com/jquery-3.7.1.min.js"
         integrity="sha256-/JqT3SQfawRcv/BIHPThkBvs0OEvtFFmqPF/lYI/Cxo=" crossorigin="anonymous"></script>
@@ -207,6 +222,7 @@ require("utilities/recent_text_translation.php");
         crossorigin="anonymous" referrerpolicy="no-referrer"></script>
     <script src="scripts/index.js"></script>
     <script src="scripts/translation_process2.js"></script>
+    <script src="scripts/newindex.js"></script>
     <!-- <script src="scripts/delete.js"></script> -->
